refactor level seed test to avoid initialization error as Firebase requires a private-key json file to start it
id rather fix this problem but i wasnt able to
